feat:artifactory服务增加通用的静态文件上传接口 #8292
