+++ src/backend/ci/core/artifactory/biz-artifactory/src/main/kotlin/com/tencent/devops/artifactory/resources/UserBkRepoStaticResourceImpl.kt
@@ -27,88 +27,38 @@
 
 package com.tencent.devops.artifactory.resources
 
-import com.tencent.devops.artifactory.api.user.UserFileResource
-import com.tencent.devops.artifactory.constant.ArtifactoryMessageCode
+import com.tencent.devops.artifactory.api.user.UserBkRepoStaticResource
 import com.tencent.devops.artifactory.pojo.enums.FileChannelTypeEnum
 import com.tencent.devops.artifactory.pojo.enums.FileTypeEnum
 import com.tencent.devops.artifactory.service.ArchiveFileService
-import com.tencent.devops.common.api.exception.ErrorCodeException
-import com.tencent.devops.common.api.exception.ParamBlankException
-import com.tencent.devops.common.api.exception.PermissionForbiddenException
+import com.tencent.devops.artifactory.util.BkRepoUtils.BKREPO_STATIC_PROJECT_ID
 import com.tencent.devops.common.api.pojo.Result
+import com.tencent.devops.common.api.util.UUIDUtil
 import com.tencent.devops.common.web.RestResource
 import org.glassfish.jersey.media.multipart.FormDataContentDisposition
 import org.springframework.beans.factory.annotation.Autowired
 import java.io.InputStream
-import javax.servlet.http.HttpServletResponse
 
-@Suppress("ThrowsCount")
 @RestResource
-class UserFileResourceImpl @Autowired constructor(
+class UserBkRepoStaticResourceImpl @Autowired constructor(
     private val archiveFileService: ArchiveFileService
-) : UserFileResource {
-
-    private fun checkParam(userId: String, projectId: String, path: String) {
-        if (userId.isBlank()) {
-            throw ParamBlankException("Invalid userId")
-        }
-        if (projectId.isBlank()) {
-            throw ParamBlankException("Invalid projectId")
-        }
-        if (path.isBlank()) {
-            throw ErrorCodeException(errorCode = ArtifactoryMessageCode.INVALID_CUSTOM_ARTIFACTORY_PATH)
-        }
-    }
-
-    override fun uploadToPath(
-        userId: String,
-        projectId: String,
-        path: String,
-        inputStream: InputStream,
-        disposition: FormDataContentDisposition
-    ): Result<String?> {
-        checkParam(userId, projectId, path)
-        val url = archiveFileService.uploadFile(
-            userId = userId,
-            inputStream = inputStream,
-            disposition = disposition,
-            projectId = projectId,
-            filePath = path,
-            fileType = FileTypeEnum.BK_CUSTOM,
-            fileChannelType = FileChannelTypeEnum.WEB_SHOW
-        )
-        return Result(url)
-    }
-
-    override fun downloadFileToLocal(userId: String, filePath: String, response: HttpServletResponse) {
-        val validateResult = archiveFileService.validateUserDownloadFilePermission(userId, filePath)
-        if (!validateResult) {
-            throw PermissionForbiddenException("no permission")
-        }
-        return archiveFileService.downloadFileToLocal(userId, filePath, response)
-    }
-
-    override fun downloadFile(userId: String, filePath: String, logo: Boolean?, response: HttpServletResponse) {
-        val validateResult = archiveFileService.validateUserDownloadFilePermission(userId, filePath)
-        if (!validateResult) {
-            throw PermissionForbiddenException("no permission")
-        }
-        archiveFileService.downloadFile(userId, filePath, response, logo)
-    }
-
-    override fun downloadFileExt(userId: String, filePath: String, logo: Boolean?, response: HttpServletResponse) {
-        downloadFile(userId, filePath, logo, response)
-    }
+) : UserBkRepoStaticResource {
 
     override fun uploadStaticFile(
         userId: String,
         inputStream: InputStream,
         disposition: FormDataContentDisposition
     ): Result<String?> {
+        val fileName = disposition.fileName
+        val index = fileName.lastIndexOf(".")
+        val fileSuffix = fileName.substring(index + 1)
+        val filePath = "file/$fileSuffix/${UUIDUtil.generate()}.$fileSuffix"
         val url = archiveFileService.uploadFile(
             userId = userId,
             inputStream = inputStream,
             disposition = disposition,
+            projectId = BKREPO_STATIC_PROJECT_ID,
+            filePath = filePath,
             fileType = FileTypeEnum.BK_STATIC,
             fileChannelType = FileChannelTypeEnum.WEB_SHOW
         )
