+++ support-files/templates/#etc#ci#common.yml
@@ -314,3 +314,10 @@ sharding:
 
 cluster:
   tag: __BK_CI_CLUSTER_TAG__
+
+# bkrepo仓库配置  
+bkrepo:
+  bkrepoApiUrl: __BK_HTTP_SCHEMA__://__BK_REPO_HOST__/api/generic
+  staticRepoPrefixUrl: __BK_HTTP_SCHEMA__://__BK_REPO_HOST__/generic/{0}/{1}
+  staticUserName: __BK_REPO_USERNAME__
+  staticPassword: __BK_REPO_PASSWORD__
